requirements.txt
tqdm
pydub
indic-nlp-library
inflect